AUTOMATED GUIDED ROBOT
It’s an autonomous car which follows a visual line embedded on the floor or ceiling (usually, a black line on a white surface). The car will transfer weights along this visual line. The user has to do all the loading and unloading. The car will start driving once its switch is turned on, carry the weights through a desired path and stop after it will reach its destination.