Create Daily Meal Expense Calculation
This project is for people or students living in mess/dorm. In a mess/dorm daily meal is noted by a manager among them manually in a page. Besides daily meal, daily bazaars as well as other utility bills are added to it. But at the end of the month its well messed up to calculate meal rate by counting individual meal.